initial commit of tree-sitter grammar.  some advanced types still in progress.
this may be based heavily upon a certain other schema language i've
previously worked on, which showed surprisingly good traction even in
days before it had any tooling at all.  There's a few critical fixes
and refinements to be made along the way, but it seems like a good
basis.  Certainly at least for concepts.

tree-sitter generates a LOT of files out of the gate.  I don't know
what the majority of them are for, tbqh.  I'm committing the ones that
I definitely recognize some authorship went into, and a few of the
git-related ones or markers of what other files are generated,
but I'm resisting actually doing "git add ." on this dir because the
result seems like it would be... well, slightly crazy.
(In particular, it sure looks to me like the whole "src" dir is...
ironically, completely generated.  Man.  Words have no meaning.)
This set of what's committed may evolve; I'm still new to tree-sitter.