indexmap dep; sketch some types.
quick debate ahead about how much I want to do of this...
in which language.

wasm experiments incoming before going much further.

(i definitely want to be able to have a good interface with
web tech; that means EITHER wasm has to kind of rock, OR
i probably actually want to write a lot of stuff in typescript.)