hello world of wasm, and an html file integrating it.
That was... a bit of a journey.  There are a lot of ways to do this,
and it seems that most of them ask a person to do a bunch of things
that... are not necessary.

What I've got here gets the number of serial document loads down to
a minimum: there's the html page, and it loads the wasm.
(I've seen other demos involve two other JS loads in the middle,
and... why.  A single script tag of "module" type readily avoids that.)

CORS is a pain in the ass.  I understand why browsers changed to
treating local files as "opaque origin", but I'm absolutely gobsmacked
that there's (afaik) no alternative configuration options that let
one work safely and without launching a whole webserver process.
Feels like the whole web platform family dropped a big, big ball there.
Anyway: workarounds in the comments in the files.

Not yet demonstrated: if we can call all the way through to the
tree-sitter bindings on the rust side, and get all *that* into wasm
without clunks.  Sure hope that works.