continuing research on getting wasm including treesitter.
It's not flying right yet.  It's still getting hung up on clang
looking for an stdio.h file, because the tree-sitter crate has
an alloc.h file that's asking for one, and I think that's... not the
place we want to be in, somehow.

I know I'm not the first one to come this way, but the internet is
strew with corpses (metaphorically speaking) and precious few
beckoning hands that seem to have actually reached their destinations.