okay, one critical step forward on wasm but still no victory
Disable default features on tree-sitter crate: the "std" dep is enabled
by default.

Guess that incantation should be "obvious" to a rust expert but,
A) I'm not one, and
B) gosh I wish someone would put these in the readme of something.

But this still leaves us with more errors:

> src/tree_sitter/parser.h:10:10: fatal error: 'stdlib.h' file not found

So now it's in the code that tree-sitter CLI has generated for me.
Which is at a whole different phase that comes earlier than this
rusty stuff.  So... Okay.  More flags/configs somewhere else, then?

The journey continues.

AMEND: false alarm, this in fact fixed nothing...?  I was just misled
by cargo's nondeterministic order of work and early exits for this
category of error causing different output during one run.

I think.

THE JOURNEY.  CONTINUES.