snapshot of more attempts to wasm; this compiles, and runs, but uncommenting actual use of tree-sitter still fails in the browser.
I sorta can't believe any of this is normal, but this is all the
incantations and filler material I ended up with to try to get
wasm-pack to compile this monster.  It now compiles (but still
doesn't _work_).

There's a whole bunch of stubs pretending to be a libc, for the
tree-sitter library and generated parser code to refer to.  Some of it
even gets implementations; a lot of it doesn't (and that hasn't
produced problems... Yet?).

The wasm feature flags on tree-sitter turned out NOT to be the way, and
in fact made things much worse.  I... don't even have a comment to make
at this point.

I'm frankly hammering things at each other without much clue.

The remaining issue (that I can see) is now at runtime in the browser.
I can call a function from my fun crate with all its spicy dependencies.
I can even return a string and print it.  But the instant I call the
first tree-sitter function... That dies: and the browser console says
something about missing an "env" module.

A subject about which there appear to be literally hundreds of issues
that also refer to wasm-pack.  Possibly thousands of comments.  All of
them seeming confused, with eschatological guesses galore from dozens
of folks, with occasionally rare stellations of "works for me", but
none of those with reproducible instructions nor any actual logical
explanations of _why_ anything in sight is working or not.   And these
sorts of issues can be found going back half a decade and yet also new
ones still showing up fresh within the last month.  Oooooookay.

So this is going to be a doozy, whatever it is.