snapshot of wasm almost livable (maybe?)
Turns out the complaint about a missing "env" module has nothing to do
with what the word "env" probably leads you to expect.

The next step towards understanding was to turn the wasm into text.
There, one can see that the "env" stuff is apparently just a
placeholder name this bindgen tooling is generating for when a symbol
isn't found provided.  Okay.

So one "env" ref is getting generated for each of the printf symbols.

Which are the things I declined to provide because they need variadics.
Which I avoided that's behind a feature flag in rust, so needs nightly
(see https://github.com/rust-lang/rust/issues/44930).  Okay.

Well, aside from the fact the discoverability of this sucks, and the
default module name that this bindgen tooling picks for this is
actively misleading, and that the suckfulness of this discoverablity
really fucking matters because ANY missing symbols would lead a wide
variety of people and projects to this exact common sticking point...

Options?  Not sure.

Here, I'm trying to provide the symbols, but incorrectly, and that...
actually, it seems to work?  And as you'll notice, these stubs were
panicking unconditionally if called _anyway_, so they're evidentally
not consequential to functioning of our code corpus.

So if that is actually working, not hiding-more-trouble-under-a-rug
that just hasn't popped out to jumpscare me again yet... okay, then
we did a double barrel roll and dodged the need for rust nightly
yet again, just barely.  Cool.  If true.  Moving on.

So now the browser console is dropping me this:
"Uncaught InternalError: too much recursion".
And points at a line of wasm that is doing... `call $func65`.

Okay.  I guess now I would like sourcemaps.
I do not currently have sourcemaps.
*sigh*