okay, at this point I have to suspect something in this wasm toolchain is just trolling me.
The "func65" I lamented in the last commit turns out to have been
the "memset" function; fortunately, when they're exported, you can
read that easily at the top of the text form wasm
(no sourcemap needed).

But how that function compiled to an infinite recursion I don't know.

This for-loop manually unrolled into the function body is safe.
This does not make sense to me, but is apparently true.

So now, in the current committed state, I get ANOTHER
appearance of "Internal Error: too much recursion"... and this time
it's from "memcmp".

And again, I do not see what the problem here should be.

Looking at this logically, my impression is some part of the compiler
toolchain here is being just flat out completely wrong.
It's showing the body of the function exported as "memcmp" as:

  local.get $var0
  local.get $var1
  local.get $var2
  call $func82

... where func82 is itself again.

I'm sorry.  What.

There is no way that's a sane compilation of the source of memcmp.
That's not even close.
That's not even *trying*.
It's a stub that reiterates its argument list.
This is just completely wrong.

I don't know what to do with this.