more thuggish impl of memcpy that can actually be crosscompiled.
(I misread critically in the previous commit's log of trouble.
It was memcpy, not memcmp, that was getting the utterly bogus body.)

Which, fair enough.  `std::ptr::copy_nonoverlapping`, the thing
previously called here, dives itself right into some rust
intrinsics shenanigans, and I can see how that'd be... hard to follow.

Okay, so now?

"Uncaught RuntimeError: unreachable executed".

Which generally means something hit a panic somewhere.

Unfortunate the stack trace attached to this in the console log is
nearly completely useless, because it taps out in the js file.
Meaning: it gives us the top function (in this case, "greet"), but
not a single clue beyond that.  Okay.  Cool.

Gotta say.  I do not know how anyone can possibly get anything done
with wasm.  This state of affairs around all this tooling and all
these error messages is extermely Not Good.  I'm real close to throwing
in the towel here, because even if I win all these current battles,
this is looking like it's going to be wandering through some kind of
endless punching-bag testing facility when getting into actual dev, too.