console_error_panic_hook needs setup; do so.
Confirmed that if I panic in rust code, this is now neato.

However, whatever my issues are with tree-sitter, they're... not
covered by this.

I'm still running totally blind.  I have no idea how to debug or
understand the 'unreachable' that's getting hit in this simple act of
trying to initialize a dang parser object.