much more legible and debuggable stuff with the --dev flag to wasm-pack.
Now, in the text form of the wasm, it's clear: there's code being
generated which is like the following:

  (func $signature_mismatch:ts_parser_set_logger (;704;) (param $var0 i32) (param $var1 i32) (param $var2 i32)
      unreachable
  )

... and yeah, okay.  I guess that makes a statement.

That's the first function we hit, but it's not the only one.
(More code uncommented in this diff as part of showing that;
the ts_parser_set_logger function is the only one that Parser::new ends
up hitting upon, but there's plenty more scattered around.)

(The fprintf stubs also show up as mismatches, but for now I'm still
going to keep fingers cross that that's going to remain irrelevant
if they're not actually encountered.)

So.  Okay.  It looks like... dang near every C function, really.

Ahhmm.